OdinOptimal/python_backend/requirements.txt
gunicorn==21.2.0
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.10
requests==2.31.0
numpy>=2.0.0,<2.4.0
scipy>=1.14.0,<1.15.0
openai>=1.30.0,<2.0.0
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
celery==5.3.4
redis==5.0.1
websockets==12.0
pytest==7.4.3
pytest-flask==1.3.0